refactor: reorganize scheduler into packages
