+++ internal/handler/scheduler/types.go
@@ -1,0 +1,8 @@
+package scheduler
+
+// ErrorResponse represents an error response
+type ErrorResponse struct {
+	Error   string `json:"error"`
+	Message string `json:"message"`
+	Code    int    `json:"code"`
+}
